modified translation speed to adapt to zoom level
translation/panning speed (in terms of axis units per unit time) is the same regardless of zoom level thus the panning speed will  be very slow at the beginning of the animation and very fast at the end, causing the animation to be less smooth. fixed in this commit
there is a another issue to do with the zoom pivot being fixed at (0,0) pixels on the chart. in the case where the final position of the view is to the right of the initial position, this causes the animation to veer left when zooming before readjusting rightwards as the translation/panning takes over. yet to find solution